# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Privacy Friendly Secret Santa Service",
        "title": "Krampus API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/draw": {
            "post": {
                "description": "Create a drawing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drawing"
                ],
                "summary": "Create a drawing",
                "parameters": [
                    {
                        "description": "Drawing data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drawing.Drawing"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/drawing.Drawing"
                        }
                    }
                }
            }
        },
        "/draw/{uuid}": {
            "get": {
                "description": "Get a drawing by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drawing"
                ],
                "summary": "Show a drawing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drawing UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/drawing.Drawing"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a drawing by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drawing"
                ],
                "summary": "Update a drawing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drawing UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Drawing data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drawing.Drawing"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/drawing.Drawing"
                        }
                    }
                }
            }
        },
        "/draw/{uuid}/member": {
            "post": {
                "description": "Create a member for a drawing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drawing"
                ],
                "summary": "Member create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drawing UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Member data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drawing.DrawingMember"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/drawing.DrawingMember"
                        }
                    }
                }
            }
        },
        "/draw/{uuid}/member/{memberUuid}": {
            "get": {
                "description": "Get a member by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Show a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drawing UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member UUID",
                        "name": "memberUuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/drawing.DrawingMember"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.defaultResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "drawing.Drawing": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/drawing.DrawingConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/drawing.DrawingGroup"
                    }
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/drawing.DrawingMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "visited_at": {
                    "type": "string"
                }
            }
        },
        "drawing.DrawingConfig": {
            "type": "object",
            "properties": {
                "chained": {
                    "type": "boolean"
                },
                "prevent_group_members": {
                    "type": "boolean"
                },
                "price": {
                    "type": "integer"
                },
                "renew_link": {
                    "type": "boolean"
                }
            }
        },
        "drawing.DrawingGroup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "drawing_uuid": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/drawing.DrawingMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "drawing.DrawingMember": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "drawing_group_uuid": {
                    "type": "string"
                },
                "drawing_uuid": {
                    "type": "string"
                },
                "drawn_member": {
                    "$ref": "#/definitions/drawing.DrawingMember"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "wishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/drawing.DrawingMemberWish"
                    }
                }
            }
        },
        "drawing.DrawingMemberWish": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "drawing_member_uuid": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "main.defaultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}